Move all MLS tests into mls subtree
